requirements.txt
numpy
pandas
statsmodels
pystan
plotly
cython
ipykernel